Allows failed controller to reconnect to a robot (#3005)
Remove the dead controller connection as usual, then recreate a new one.
If only one external controller in the world, this one is restarted.
Otherwise, it use the name of the robot to know which dead controller should be kill.
Otherwise, if no name is provided (or no robot matching), the code execute as usual (scanning for free spot.